refactor: densify connectivity dev tests, ratchet package LOC to 15900 (#4246)
Consolidate connectivity snapshot/parity/determinism pins into shared
fixtures and merge runners into connectivity_dev_targets_test; extract
naval mission draft scenarios to support. Package test LOC 16095 → 15801.

Refs #4246
